Restructured implementation of FunctionTree and Calc trait
